Added new volunteer form in ajax style

+++ db/migrate/20210329210651_create_volunteers.rb
@@ -2,7 +2,7 @@ class CreateVolunteers < ActiveRecord::Migration[6.0]
   def change
     create_table :volunteers do |t|
       t.string :name
-      t.integer :contactnumber
+      t.bigint :contactnumber
       t.string :addressone
       t.string :addresstwo
       t.string :email
